requirements.txt
streamlit
langchain
langchain-groq
sentence-transformers
faiss-cpu
google-search-results
numpy